+++ packages/cli/doc-plugin-auto-sidebar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/doc-plugin-auto-sidebar
 
+## 2.10.0
+
+### Patch Changes
+
+- Updated dependencies [c413d79]
+- Updated dependencies [0da32d0]
+- Updated dependencies [fbefa7e]
+- Updated dependencies [4d54233]
+- Updated dependencies [6db4864]
+  - @modern-js/doc-tools@2.10.0
+  - @modern-js/utils@2.10.0
+
 ## 2.9.0
 
 ### Patch Changes