+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @modern-js/plugin-bff
 
+## 2.10.0
+
+### Patch Changes
+
+- 0da32d0: chore: upgrade jest and puppeteer
+  chore: 升级 jest 和 puppeteer 到 latest
+- 0d9962b: fix: add types field in package.json
+  fix: 添加 package.json 中的 types 字段
+- fbefa7e: chore(deps): bump webpack from 5.75.0 to 5.76.2
+
+  chore(deps): 将 webpack 从 5.75.0 升级至 5.76.2
+
+- Updated dependencies [0da32d0]
+- Updated dependencies [0d9962b]
+- Updated dependencies [fbefa7e]
+- Updated dependencies [4d54233]
+- Updated dependencies [6db4864]
+  - @modern-js/create-request@2.10.0
+  - @modern-js/bff-core@2.10.0
+  - @modern-js/utils@2.10.0
+  - @modern-js/server-utils@2.10.0
+
 ## 2.9.0
 
 ### Patch Changes