+++ packages/toolkit/e2e/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/e2e
 
+## 2.10.0
+
+### Patch Changes
+
+- Updated dependencies [0da32d0]
+- Updated dependencies [fbefa7e]
+- Updated dependencies [4d54233]
+- Updated dependencies [6db4864]
+  - @modern-js/utils@2.10.0
+
 ## 2.9.0
 
 ### Patch Changes