+++ scripts/lint-package-json/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scripts/lint-package-json
 
+## 2.10.0
+
 ## 2.9.0
 
 ## 2.8.0